requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
openai
python-dotenv
pydantic